Various fixes for PR comments
