+++ pallets/stateful-storage/src/tests/itemized_operations_tests.rs
@@ -1,6 +1,6 @@
 use crate::{
 	stateful_child_tree::StatefulChildTree,
-	test_common::{constants::*, test_constants::*},
+	test_common::{constants::*, test_utility::*},
 	tests::mock::*,
 	types::*,
 	Config, Error,
